Optimized GIF; reorganized code, directories
Added a delay at the end of the saved GIF.

Switched sample skeleton image to a picture
of a woman instead of the pair of elephants.

Added input and output directories for better
organization.

Updated sample output GIF.

Updated file input section of code for ease
of use.

Updated .gitignore